+++ packages/components/src/Chips/InternalChipDismissible/InternalChipDismissible.module.css.d.ts
@@ -2,7 +2,6 @@ declare const styles: {
   readonly "wrapper": string;
   readonly "input": string;
   readonly "menu": string;
-  readonly "menuList": string;
   readonly "menuListOption": string;
   readonly "loadingIndicator": string;
   readonly "activeOption": string;